MyMalloc
Implement our own version of malloc() and free(). The advantages of our
program is that it lets us detect common dynamic memory errors. We
implemented a static
char array for the dynamic memory.
The static char array contains our dynamic memory however we also have
an array that
contains void pointers to memory access. We initialized all the cells in
our array to 0 in order to
keep track of memory which has been malloc’ed for error detection.
When malloc() is used space a specific amount of space is given based on
the request.
MemAccess contains a header that lets us deal with how much space is
used. The pointer of the
MemAccess struct is located in the first index of the MemAccess array.
When free() is used, the
memory size of MemAccess is taken off from the address which the pointer
is stored to and is
then validified with our MemAccess array.